test(lfm): drop the batched-format tests, keep the global leg as its own module
The batched proof format is retired, so every test whose SUBJECT is that
format goes. Deletions, by file:

- `aggregator_tests.rs` (2,664 lines, 10 tests) — the module is typed end to
  end on `BatchedMultiProof`. Its GLOBAL leg is not: the cross-epoch global
  memory proof is per-table under either format, so `RealGlobal`,
  `real_global`, `GlobalTableArenas`, `global_verifier_program`,
  `global_arena_words` and `the_global_verifier_leg_runs_and_rejects_tampers`
  move verbatim into `per_table_aggregator_tests.rs`. One substitution: the
  fixture bundle now comes from `prove_continuation` instead of
  `prove_continuation_batched` — identical signature, and `real_global` already
  took a plain `ContinuationProof`.
- `group_leaf_tests.rs` (295 lines, 3 tests) — the mixed-height group leaf.
- `epoch_tests.rs` — the batched harness and its 7 tests, plus
  `the_real_block_base_epoch_ab`, whose subject was per-table VS batched and
  which cannot be a one-armed A/B. That test had absorbed `real_epoch_with`'s
  doc block (no blank line between them); deleting it returns the doc to
  `real_epoch_with`.
- `wrap_tests.rs` — the batched wrap section, its 3 tests and the P2 batched
  block driver. That driver's doc had likewise absorbed
  `the_real_blocks_page_census`'s; the census doc is reattached.
- `epoch_verify_tests.rs` — the batched leg-input builders.
- `machine_tests.rs` — the three batched LFM epoch/wire tests.
- `algebraic_commit.rs` — `the_mixed_group_leaf_equals_the_hosts_group_hash`.
  A group leaf over several matrices is the mixed-height construction; the
  one-matrix case it degenerates to is `the_emitted_leaf_and_parent_equal_the_hosts_backend`'s
  subject and stays gated. `StarkHash::Batched` is the batch-commitment leaf
  backend, not the format, and is untouched.

Two tests are EDITED rather than deleted, because their per-table half is the
claim and the batched half was an arm of it:

- `blake3_chip_tests::chunked_blake3_proves_and_verifies` loses its batched
  arm; the per-table prove/verify and the host-chain digest equality stay.
- `algebraic_transcript::phase_a_absorbs_derive_the_hosts_shared_pair` loses
  the shape histogram, both sides of it. What the gate is actually for
  survives intact: that a preprocessed root reaching the transcript as program
  TEXT (`append_const_bytes`) and as proof-carried cells
  (`RootCells::constant` + `absorb_misaligned`) derive the same challenge,
  i.e. that `bytes_to_cell` and `commitment_to_digest` agree. That property is
  Phase A's on either format and is invisible under a byte hash.
