+++ prover/src/lfm/epoch_tests.rs
@@ -25,9 +25,6 @@
 //! [`the_legs_consume_the_replayed_challenges`]'s job, and the whole-epoch
 //! composition (25 sub-proofs behind one statement) is not built here.
 
-use stark::batched::proof::{BatchedMultiProof, BatchedProveStats};
-use stark::batched::shape::{EpochFriParams, EpochShape};
-use stark::batched::verifier::EpochChallenges;
 use stark::config::Commitment;
 use stark::proof::stark::MultiProof;
 use stark::proof::view::StarkProofView;
@@ -722,125 +719,15 @@ impl EpochInputs {
 /// is the fibonacci fixture unless a measurement run overrode it — so two runs
 /// at different options stay comparable, and assembly ledger entry 10 still
 /// holds: the trace-length profile travels with every number.
-/// ★ THE BASE-LAYER A/B — the real block's epoch 0 proved per-table vs
-/// BATCHED-MMCS, one arm per process.
-///
-/// `AB_MODE` selects the arm (`per_table` | `batched`); `LAMBDA_VM_RESIDENCY`
-/// moves BOTH arms through the same lever, so a residency difference between
-/// them cannot be an artifact of two code paths reading two knobs. Peak anon
-/// is a process-lifetime high-water mark, measured by the harness around the
-/// process — two arms sharing a process would each report the larger of the
-/// two and the comparison would be vacuous.
-///
-/// The epoch construction IS the census harness's ([`EpochFront`], the same
-/// call [`real_epoch_from`] makes): same executor slice, same traces, same
-/// L2G bookend, same statement-seeded transcript. Only the prove call
-/// differs.
-///
-/// ⚠ NEITHER arm verifies here, deliberately. Each arm's construction is
-/// production-accepted by its own gate elsewhere — the per-table one every
-/// time `the_real_block_epoch_wraps` runs, the batched one by
-/// [`a_batched_vm_epoch_host_verifies_end_to_end`] (per-table preprocessed
-/// binding made `multi_verify_batched` a complete verification for the VM
-/// AIR set too) — and a verify inside a memory instrument would smear its
-/// own footprint over the number being measured. This instrument measures
-/// the PROVE.
-#[test]
-#[ignore]
-fn the_real_block_base_epoch_ab() {
-    use stark::prover::IsStarkProver;
-
-    for var in ["LFM_CENSUS_ELF", "LFM_CENSUS_INPUT"] {
-        assert!(
-            std::env::var(var).is_ok(),
-            "{var} must name a file: this A/B measures a REAL block epoch"
-        );
-    }
-    let mode = std::env::var("AB_MODE").expect("AB_MODE must be per_table or batched");
-    let residency = match std::env::var("LAMBDA_VM_RESIDENCY").as_deref() {
-        Ok("recompute") => stark::residency_mode::ResidencyMode::RecomputeLde,
-        _ => stark::residency_mode::ResidencyMode::Retain,
-    };
-
-    let inputs = EpochInputs::from_env();
-    let opts = crate::recursion::Preset::Blowup4.options();
-    let mut inner = opts;
-    if let Ok(v) = std::env::var("LFM_WRAP_QUERIES") {
-        inner.fri_number_of_queries = v.parse().expect("LFM_WRAP_QUERIES must be an integer");
-    }
-    let opts = inner;
-    println!(
-        "★ BASE A/B ARM: mode={mode} residency={residency:?}  guest {}, \
-         2^{} cycles/epoch, blowup {} / {} queries",
-        inputs.label, inputs.epoch_log2, opts.blowup_factor, opts.fri_number_of_queries,
-    );
-
-    let mut front = EpochFront::build(opts, inputs);
-    let mut transcript = front.seed();
-    let pairs = front.pairs();
-
-    // ---- THE MEASURED PROVE. Everything above is identical shared setup.
-    let t = std::time::Instant::now();
-    match mode.as_str() {
-        "per_table" => {
-            let proof = crate::hash_pin::BlockProver::<Gl, Ext3, ()>::multi_prove(
-                pairs,
-                &mut transcript,
-                #[cfg(feature = "disk-spill")]
-                stark::storage_mode::StorageMode::Ram,
-                residency,
-            )
-            .expect("the epoch must prove");
-            let prove_secs = t.elapsed().as_secs_f64();
-            let size = rkyv::to_bytes::<rkyv::rancor::Error>(&proof)
-                .expect("the epoch proof must serialize")
-                .len();
-            println!(
-                "★ BASE A/B RESULT mode=per_table PROVE_SECS={prove_secs:.2} \
-                 SUB_PROOFS={} PROOF_BYTES={size}",
-                proof.proofs.len(),
-            );
-        }
-        "batched" => {
-            let (proof, stats) = stark::batched::prover::multi_prove_batched::<
-                Gl,
-                Ext3,
-                (),
-                crate::hash_pin::BlockStarkHash,
-                crate::hash_pin::BlockProver<Gl, Ext3, ()>,
-            >(
-                pairs,
-                &mut transcript,
-                #[cfg(feature = "disk-spill")]
-                stark::storage_mode::StorageMode::Ram,
-                residency,
-            )
-            .expect("the batched epoch must prove");
-            let prove_secs = t.elapsed().as_secs_f64();
-            println!(
-                "★ BASE A/B RESULT mode=batched PROVE_SECS={prove_secs:.2} \
-                 TABLES={} QUERIES={} FRI_LAYERS={} PREP_TABLES={}",
-                proof.tables.len(),
-                proof.queries.len(),
-                proof.fri_layer_roots.len(),
-                proof.queries.first().map_or(0, |q| q.prep.len()),
-            );
-            println!("   BATCHED_STATS {stats:?}");
-        }
-        other => panic!("AB_MODE must be per_table or batched, not {other}"),
-    }
-}
-
 pub(super) fn real_epoch_with(opts: crate::ProofOptions) -> RealEpoch {
     real_epoch_from(opts, EpochInputs::from_env())
 }
 
 /// The statement-seeded transcript every prover and every verifier of one
-/// epoch starts from. One function rather than per-harness closures so the
-/// per-table and batched harnesses CANNOT drift on the absorb — a drift here
-/// would fail neither harness's own gate; it would just make their proofs
-/// answer different statements, which is exactly the failure a per-table vs
-/// batched comparison cannot detect from inside.
+/// epoch starts from. One function rather than per-harness closures so no two
+/// harnesses can drift on the absorb — a drift here would fail neither
+/// harness's own gate; it would just make their proofs answer different
+/// statements, which is a failure no harness can detect from inside.
 pub(super) fn epoch_seed(
     epoch_label: u64,
     elf_bytes: &[u8],
@@ -863,13 +750,11 @@ pub(super) fn epoch_seed(
     t
 }
 
-/// The census-env epoch-0 construction, shared BY STRUCTURE between the
-/// per-table harness ([`real_epoch_from`]), the base-layer A/B and the
-/// batched harness ([`real_batched_epoch_from`]): same executor slice, same
-/// traces, same L2G bookend, same statement-seeded transcript. The premise of
-/// every per-table vs batched comparison this file hosts is "same epoch,
-/// different prove", and sharing this front is what makes the premise
-/// structural rather than by-inspection.
+/// The census-env epoch-0 construction the per-table harness
+/// ([`real_epoch_from`]) is built on: same executor slice, same traces, same
+/// L2G bookend, same statement-seeded transcript. Sharing this front is what
+/// makes "same epoch, different prove" structural rather than by-inspection
+/// for any two runs this file hosts.
 ///
 /// Only epoch 0 is built: the boundary starts from genesis provenance and the
 /// label is `epoch_label(0)`, so a later epoch would need the previous one's
@@ -1492,2012 +1377,6 @@ fn the_from_proof_constructor_rejects_a_tampered_bundle() {
     println!(
         "★ P1 GATE C: reg_fini corruption rejected by the constructor; \
          l2g_root corruption rejected by the bundle verifier"
-    );
-}
-
-/// The batched-path analogue of [`RealEpoch`] — the host half of the M-8
-/// full-recursion campaign: the SAME construction ([`EpochFront`]), proved
-/// through `multi_prove_batched`, host-verified COMPLETELY before anything
-/// downstream reads it.
-///
-/// The struct keeps the AIR set alive because every verification —
-/// the constructor's gate and every tamper arm — rebuilds `refs()` from it,
-/// and T2's shape derivations read program shape from the AIRs rather than
-/// the proof. The traces are dropped at the end of construction: holders pay
-/// for the proof, not the epoch's tables.
-// The fields nothing reads yet are the emitter's contract (handoff T2): the
-// spine reads statement/challenges/prep provenance/register files, the legs
-// read shape/fri_params, the census reads prove_stats. Removing one because
-// it is currently unread would just re-derive it worse there.
-#[allow(dead_code)]
-pub(super) struct RealBatchedEpoch {
-    pub(super) opts: crate::ProofOptions,
-    pub(super) statement: super::statement_replay::EpochStatementShape,
-    pub(super) elf_bytes: Vec<u8>,
-    pub(super) elf_digest: [u8; 32],
-    pub(super) public_output: Vec<u8>,
-    pub(super) epoch_label: u64,
-    pub(super) table_counts: crate::TableCounts,
-    pub(super) runtime_page_ranges: Vec<crate::RuntimePageRange>,
-    airs: crate::VmAirs,
-    l2g_air: Box<dyn AIR<Field = Gl, FieldExtension = Ext3, PublicInputs = ()>>,
-    pub(super) register_init: Vec<u32>,
-    pub(super) reg_fini: Vec<u32>,
-    pub(super) pc_start: u64,
-    pub(super) expected_program_id: [u8; 32],
-    /// Per table in sub-proof order: preprocessed provenance when the AIR is
-    /// preprocessed — [`prep_source`]'s taxonomy, unchanged, because the
-    /// batched proof binds the same per-table roots the per-table path does,
-    /// so the wrap's binding machinery is the existing one. There are no
-    /// per-table main roots to pair these with: the shared mixed roots live
-    /// on the proof itself.
-    pub(super) prep_sources: Vec<Option<PrepSource>>,
-    pub(super) proof: BatchedMultiProof<Gl, Ext3, ()>,
-    pub(super) shape: EpochShape,
-    pub(super) fri_params: EpochFriParams,
-    /// Every challenge the batched transcript derives, recovered through
-    /// `replay_epoch_transcript` — the oracle T2's emitted spine
-    /// differentials against.
-    pub(super) challenges: EpochChallenges<Ext3>,
-    /// The carried commit index — `reg_init[X254_INDEX]`, same meaning as
-    /// [`RealEpoch::start_index`].
-    pub(super) start_index: u64,
-    /// The COMMIT-bus target derived from the REPLAYED shared pair, exactly
-    /// as `verify_against_batched` derives it: the batched path has no
-    /// per-table Phase A to walk.
-    pub(super) expected_bus_balance: FEE,
-    pub(super) prove_stats: BatchedProveStats,
-}
-
-impl RealBatchedEpoch {
-    /// [`epoch_seed`] over this epoch's statement, with the claimed output
-    /// substitutable so a tamper arm can ask the question it means: "does
-    /// THIS proof answer for THAT output?".
-    fn seed_for(&self, public_output: &[u8]) -> crate::hash_pin::BlockTranscript {
-        epoch_seed(
-            self.epoch_label,
-            &self.elf_bytes,
-            public_output,
-            &self.table_counts,
-            &self.runtime_page_ranges,
-            self.opts.fri_final_poly_log_degree,
-        )
-    }
-
-    /// The AIR set in sub-proof order — the VM tables then the L2G bookend,
-    /// the same order [`EpochFront::pairs`] proved in.
-    pub(super) fn refs(
-        &self,
-    ) -> Vec<&dyn AIR<Field = Gl, FieldExtension = Ext3, PublicInputs = ()>> {
-        let mut r = self.airs.air_refs();
-        r.push(&*self.l2g_air);
-        r
-    }
-
-    /// The COMPLETE host verification of `proof` against this epoch's AIR
-    /// set and the given claimed output, mirroring `verify_against_batched`:
-    /// the challenges replayed on a fork of the statement seed, the expected
-    /// COMMIT-bus balance from the replayed shared pair, then
-    /// `multi_verify_batched`. `false` on any tamper; never panics on proof
-    /// data.
-    pub(super) fn host_verifies_for(
-        &self,
-        proof: &BatchedMultiProof<Gl, Ext3, ()>,
-        claimed_output: &[u8],
-    ) -> bool {
-        // One code path for both formats: the carve configuration rides in
-        // the shape this epoch was replayed under (None ≡ uncarved).
-        let carved = self.shape.carved_main.map(|c| c.table);
-        let refs = self.refs();
-        let mut replay = self.seed_for(claimed_output);
-        let Some((_, _, challenges)) = stark::batched::verifier::replay_epoch_transcript_carved(
-            &refs,
-            proof,
-            &mut replay,
-            carved,
-        ) else {
-            return false;
-        };
-        let [z, alpha] = challenges.lookup.as_slice() else {
-            return false;
-        };
-        let Some(expected) =
-            crate::compute_commit_bus_offset(claimed_output, self.start_index, z, alpha)
-        else {
-            return false;
-        };
-        let mut transcript = self.seed_for(claimed_output);
-        stark::batched::verifier::multi_verify_batched_carved::<
-            Gl,
-            Ext3,
-            (),
-            crate::hash_pin::BlockStarkHash,
-            crate::hash_pin::BlockVerifier<Gl, Ext3, ()>,
-            _,
-        >(&refs, proof, &mut transcript, &expected, carved)
-    }
-
-    /// [`RealBatchedEpoch::host_verifies_for`] at this epoch's own output.
-    pub(super) fn host_verifies(&self, proof: &BatchedMultiProof<Gl, Ext3, ()>) -> bool {
-        self.host_verifies_for(proof, &self.public_output)
-    }
-}
-
-pub(super) fn real_batched_epoch_with(opts: crate::ProofOptions) -> RealBatchedEpoch {
-    real_batched_epoch_from(opts, EpochInputs::from_env())
-}
-
-/// [`real_epoch_from`]'s batched sibling. Panics — loudly, this is a harness
-/// — if the proof does not host-verify: nothing downstream may read an epoch
-/// production would reject.
-pub(super) fn real_batched_epoch_from(
-    opts: crate::ProofOptions,
-    inputs: EpochInputs,
-) -> RealBatchedEpoch {
-    real_batched_epoch_from_with_carve(opts, inputs, false)
-}
-
-/// [`real_batched_epoch_from`] with the L2G table CARVED — the continuation
-/// batched format ([`crate::continuation::prove_continuation_batched`]'s
-/// per-epoch shape), for the carved emitter gates.
-pub(super) fn real_batched_epoch_carved_from(
-    opts: crate::ProofOptions,
-    inputs: EpochInputs,
-) -> RealBatchedEpoch {
-    real_batched_epoch_from_with_carve(opts, inputs, true)
-}
-
-fn real_batched_epoch_from_with_carve(
-    opts: crate::ProofOptions,
-    inputs: EpochInputs,
-    carve_l2g: bool,
-) -> RealBatchedEpoch {
-    let mut front = EpochFront::build(opts, inputs);
-
-    let (proof, prove_stats, carved_index) = {
-        let mut transcript = front.seed();
-        let t = std::time::Instant::now();
-        let pairs = front.pairs();
-        // The L2G bookend is the LAST pair — the carved table, when carving.
-        let carved_index = carve_l2g.then(|| pairs.len() - 1);
-        let (proof, stats) = stark::batched::prover::multi_prove_batched_carved::<
-            Gl,
-            Ext3,
-            (),
-            crate::hash_pin::BlockStarkHash,
-            crate::hash_pin::BlockProver<Gl, Ext3, ()>,
-        >(
-            pairs,
-            &mut transcript,
-            #[cfg(feature = "disk-spill")]
-            stark::storage_mode::StorageMode::Ram,
-            stark::residency_mode::ResidencyMode::Retain,
-            carved_index,
-        )
-        .expect("the batched epoch must prove");
-        eprintln!(
-            "batched inner epoch: {}, 2^{} cycles, {} cycles executed, \
-             {} tables, proved in {:.1}s",
-            front.guest_label,
-            front.epoch_log2,
-            front.cycles_executed,
-            proof.tables.len(),
-            t.elapsed().as_secs_f64()
-        );
-        (proof, stats, carved_index)
-    };
-
-    // The traces fed the prove; drop them here — what follows reads the PROOF.
-    let EpochFront {
-        opts,
-        elf_bytes,
-        elf,
-        airs,
-        l2g_air,
-        register_init,
-        reg_fini,
-        table_counts,
-        public_output,
-        runtime_page_ranges,
-        label,
-        decode_root,
-        ..
-    } = front;
-
-    harvest_real_batched_epoch(
-        opts,
-        elf_bytes,
-        &elf,
-        airs,
-        l2g_air,
-        register_init,
-        reg_fini,
-        table_counts,
-        public_output,
-        runtime_page_ranges,
-        label,
-        decode_root,
-        proof,
-        carved_index,
-        prove_stats,
-    )
-    .expect("the session-built batched epoch must harvest")
-}
-
-/// Everything downstream of a batched epoch's PROOF: the replay, the COMMIT
-/// target, the preprocessed provenances, the statement shape — and the
-/// complete host verification as the acceptance gate. Shared by the session
-/// harness above and the from-continuation constructor below, so the two
-/// reconstructions cannot diverge (the P1 discipline, batched).
-#[allow(clippy::too_many_arguments)]
-fn harvest_real_batched_epoch(
-    opts: crate::ProofOptions,
-    elf_bytes: Vec<u8>,
-    elf: &executor::elf::Elf,
-    airs: crate::VmAirs,
-    l2g_air: Box<dyn AIR<Field = Gl, FieldExtension = Ext3, PublicInputs = ()>>,
-    register_init: Vec<u32>,
-    reg_fini: Vec<u32>,
-    table_counts: crate::TableCounts,
-    public_output: Vec<u8>,
-    runtime_page_ranges: Vec<crate::RuntimePageRange>,
-    label: u64,
-    decode_root: Commitment,
-    proof: BatchedMultiProof<Gl, Ext3, ()>,
-    carved_index: Option<usize>,
-    prove_stats: BatchedProveStats,
-) -> Result<RealBatchedEpoch, String> {
-    use crate::tables::register;
-
-    let refs = {
-        let mut r = airs.air_refs();
-        r.push(&*l2g_air);
-        r
-    };
-
-    // ---- the replay: every challenge, and the shape both sides derive.
-    let mut replay = epoch_seed(
-        label,
-        &elf_bytes,
-        &public_output,
-        &table_counts,
-        &runtime_page_ranges,
-        opts.fri_final_poly_log_degree,
-    );
-    let Some((shape, fri_params, challenges)) =
-        stark::batched::verifier::replay_epoch_transcript_carved(
-            &refs,
-            &proof,
-            &mut replay,
-            carved_index,
-        )
-    else {
-        return Err("the batched epoch's transcript rejects: it does not replay".to_string());
-    };
-    let [z, alpha] = challenges.lookup.as_slice() else {
-        return Err("an epoch uses LogUp, so the shared pair must be exactly (z, α)".to_string());
-    };
-    let start_index = register_init[register::X254_INDEX] as u64;
-    let Some(expected) = crate::compute_commit_bus_offset(&public_output, start_index, z, alpha)
-    else {
-        return Err("the COMMIT bus target rejects: it does not compute".to_string());
-    };
-
-    let prep_sources = refs
-        .iter()
-        .map(|air| {
-            air.is_preprocessed().then(|| {
-                prep_source(
-                    air.precomputed_commitment(),
-                    &opts,
-                    elf,
-                    &register_init,
-                    &reg_fini,
-                )
-            })
-        })
-        .collect();
-    drop(refs);
-
-    let e = RealBatchedEpoch {
-        statement: super::statement_replay::EpochStatementShape {
-            public_output_len: public_output.len(),
-            table_counts: [
-                table_counts.cpu as u64,
-                table_counts.lt as u64,
-                table_counts.memw as u64,
-                table_counts.memw_aligned as u64,
-                table_counts.load as u64,
-                table_counts.mul as u64,
-                table_counts.dvrm as u64,
-                table_counts.shift as u64,
-                table_counts.branch as u64,
-                table_counts.memw_register as u64,
-                table_counts.eq as u64,
-                table_counts.bytewise as u64,
-                table_counts.store as u64,
-                table_counts.cpu32 as u64,
-                table_counts.blake3 as u64,
-            ],
-            num_private_input_pages: 0,
-            fri_final_poly_log_degree: opts.fri_final_poly_log_degree,
-            page_ranges: runtime_page_ranges
-                .iter()
-                .map(|r| (r.base, r.count))
-                .collect(),
-        },
-        elf_digest: crate::statement::elf_digest(&elf_bytes),
-        expected_program_id: crate::recursion::program_id_from_digest(
-            &crate::statement::elf_digest(&elf_bytes),
-            elf.entry_point,
-            &decode_root,
-            &[],
-        ),
-        pc_start: elf.entry_point,
-        opts,
-        elf_bytes,
-        public_output,
-        epoch_label: label,
-        table_counts,
-        runtime_page_ranges,
-        airs,
-        l2g_air,
-        register_init,
-        reg_fini,
-        prep_sources,
-        proof,
-        shape,
-        fri_params,
-        challenges,
-        start_index,
-        expected_bus_balance: expected,
-        prove_stats,
-    };
-
-    // ---- production-shaped acceptance, or nothing above describes a real
-    // epoch. `host_verifies` is the complete check — the same derivation the
-    // per-table harness's `multi_verify_views` gate plays on its side; the
-    // tamper arms in `a_batched_vm_epoch_host_verifies_end_to_end` keep it
-    // discriminating.
-    if !e.host_verifies(&e.proof) {
-        return Err("production's batched verifier rejects this epoch".to_string());
-    }
-    Ok(e)
-}
-
-/// [`RealBatchedEpoch`] for epoch `epoch_index` of an EXISTING continuation
-/// bundle proven BATCHED — the from-proof path, mirroring
-/// [`real_epoch_from_continuation`]: the AIR set and statement values from
-/// [`crate::continuation::reconstruct_epoch_airs`] (the SAME reconstruction
-/// `verify_epoch` runs), the chain position from
-/// [`crate::continuation::epoch_chain_position`], the harvest shared with the
-/// session path. The epoch's proof is the bundle's batched body, its L2G main
-/// matrix carved (always the LAST table); production's complete batched
-/// verify inside the harvest is the acceptance gate.
-pub(super) fn real_batched_epoch_from_continuation(
-    opts: &crate::ProofOptions,
-    elf_bytes: &[u8],
-    bundle: &crate::continuation::ContinuationProof,
-    epoch_index: usize,
-    decode_commitment: Option<Commitment>,
-) -> Result<RealBatchedEpoch, String> {
-    use executor::elf::Elf;
-
-    let elf = Elf::load(elf_bytes).map_err(|e| format!("the inner ELF must load: {e}"))?;
-    let position = crate::continuation::epoch_chain_position(bundle, &elf, epoch_index)
-        .map_err(|e| format!("chain position for epoch {epoch_index}: {e:?}"))?
-        .ok_or_else(|| format!("epoch {epoch_index} is out of range or the bundle is malformed"))?;
-    let view = bundle.epoch_view(epoch_index);
-    let recon = crate::continuation::reconstruct_epoch_airs(
-        &elf,
-        view,
-        &position.register_init,
-        position.is_final,
-        position.label,
-        opts,
-        decode_commitment,
-    )
-    .map_err(|e| format!("reconstructing epoch {epoch_index}: {e:?}"))?
-    .ok_or_else(|| format!("epoch {epoch_index} is structurally invalid"))?;
-    let decode_root = match decode_commitment {
-        Some(c) => c,
-        None => crate::tables::decode::commitment_from_elf(&elf, opts)
-            .map_err(|e| format!("DECODE commitment from ELF: {e}"))?,
-    };
-    let proof = view
-        .batched_proof()
-        .ok_or_else(|| {
-            format!(
-                "epoch {epoch_index} is per-table; the batched constructor reads batched bundles"
-            )
-        })?
-        .materialize()
-        .map_err(|e| format!("materializing epoch {epoch_index}'s batched proof: {e:?}"))?
-        .into_owned();
-    let carved_index = Some(proof.tables.len() - 1);
-    harvest_real_batched_epoch(
-        opts.clone(),
-        elf_bytes.to_vec(),
-        &elf,
-        recon.airs,
-        recon.l2g_air,
-        position.register_init,
-        recon.reg_fini,
-        recon.table_counts,
-        view.public_output().to_vec(),
-        recon.runtime_page_ranges,
-        position.label,
-        decode_root,
-        proof,
-        carved_index,
-        BatchedProveStats::default(),
-    )
-}
-
-/// ★ THE H2 GATE: the same construction the per-table harness proves is
-/// proved through the BATCHED path and host-verified COMPLETELY — then
-/// tampered, so acceptance is discrimination, not a verifier that stopped
-/// checking.
-///
-/// Arms: a preprocessed opening value (the per-table critical check, on the
-/// batched proof's own per-query openings), a main-round mixed opening value
-/// (the shared-tree authentication), and the claimed output (the statement +
-/// COMMIT-bus binding; the fixture's output is empty, so the arm EXTENDS it
-/// rather than moving a byte — length is bound either way).
-#[test]
-fn a_batched_vm_epoch_host_verifies_end_to_end() {
-    let e = real_batched_epoch_with(super::proof_fixture::fixture_options());
-
-    // The replayed challenge lists are per table in table order, and the
-    // proof's per-query preprocessed openings are one per preprocessed AIR —
-    // the alignments every T2 consumer will assume, asserted where the
-    // harness is built.
-    assert_eq!(e.challenges.betas.len(), e.proof.tables.len());
-    assert_eq!(e.challenges.zs.len(), e.proof.tables.len());
-    assert_eq!(e.challenges.deep_gammas.len(), e.proof.tables.len());
-    let num_preprocessed = e.prep_sources.iter().filter(|s| s.is_some()).count();
-    for q in &e.proof.queries {
-        assert_eq!(q.prep.len(), num_preprocessed);
-    }
-
-    // The arena serializers fill EXACTLY what the shape-derived closed forms
-    // declare — the asserts live inside them; called here so the discipline
-    // gates on the same proof the tampers gate on.
-    let opening = super::epoch_verify_tests::batched_opening_arena(&e);
-    let fri = super::epoch_verify_tests::batched_fri_arena(&e);
-    assert!(
-        !opening.is_empty(),
-        "every epoch opens at least the main round"
-    );
-    eprintln!(
-        "batched arenas: {} opening words, {} FRI words over {} queries",
-        opening.len(),
-        fri.len(),
-        e.proof.queries.len()
-    );
-
-    let mut tampered = e.proof.clone();
-    tampered.queries[0]
-        .prep
-        .first_mut()
-        .expect("a VM epoch has preprocessed tables")
-        .evaluations[0] += FE::one();
-    assert!(
-        !e.host_verifies(&tampered),
-        "a tampered preprocessed opening must be rejected"
-    );
-
-    let mut tampered = e.proof.clone();
-    tampered.queries[0].main.per_matrix[0].evaluations[0] += FE::one();
-    assert!(
-        !e.host_verifies(&tampered),
-        "a tampered main-round opening must be rejected"
-    );
-
-    let mut moved = e.public_output.clone();
-    match moved.first_mut() {
-        Some(byte) => *byte ^= 1,
-        None => moved.push(1),
-    }
-    assert!(
-        !e.host_verifies_for(&e.proof, &moved),
-        "a moved claimed output must be rejected"
-    );
-}
-
-/// The batched spine's program shape, host-derived from the harness — each
-/// field the value the emitter reads off the AIR set and the options, never
-/// off the proof (the OOD dims come from the proof's blocks exactly as the
-/// per-table `TableChallengeShape` takes them, blessed as program shape for
-/// the same reason: the program is emitted for one epoch shape).
-fn batched_shape_of(e: &RealBatchedEpoch) -> super::batched_epoch::BatchedEpochShape {
-    use super::batched_epoch::{BatchedEpochShape, BatchedFriShape, BatchedTableShape};
-
-    let refs = e.refs();
-    let tables: Vec<BatchedTableShape> = e
-        .proof
-        .tables
-        .iter()
-        .zip(&refs)
-        .map(|(t, air)| BatchedTableShape {
-            log2_trace_length: t.trace_length.trailing_zeros(),
-            has_contribution: air.has_aux_trace(),
-            ood_current_dims: (
-                t.trace_ood_evaluations.width,
-                t.trace_ood_evaluations.height,
-            ),
-            ood_next_dims: (
-                t.trace_ood_next_evaluations.width,
-                t.trace_ood_next_evaluations.height,
-            ),
-            num_parts: t.composition_poly_parts_ood_evaluation.len(),
-        })
-        .collect();
-    BatchedEpochShape {
-        tables,
-        heights: e.shape.heights.clone(),
-        total_widths: e.shape.total_widths(),
-        log2_blowup: e.fri_params.blowup_log,
-        coset_offset: FE::from(e.fri_params.coset_offset),
-        has_aux: !e.shape.aux.is_empty(),
-        carved_main: e.shape.carved_main.map(|c| (c.table, c.width)),
-        fri: BatchedFriShape::new(
-            &e.shape.heights,
-            e.fri_params.blowup_log,
-            e.fri_params.final_poly_log_degree,
-        ),
-        grinding_factor: e.fri_params.grinding_factor,
-        num_queries: e.fri_params.num_queries,
-    }
-}
-
-/// The batched epoch's spine program — statement, prep provenance, the
-/// attestation join, the ONE-transcript batched challenge replay, and the
-/// LogUp closure. The batched sibling of [`epoch_program`]'s spine half.
-pub(super) fn batched_epoch_program(e: &RealBatchedEpoch) -> LfmProgram {
-    batched_epoch_program_with(e, false, false)
-}
-
-/// One table's leg shapes, from the AIR set and the proof's trace lengths —
-/// the same derivations [`build_table_legs`] makes, minus the per-table view
-/// the batched proof does not have.
-///
-/// [`build_table_legs`]: super::epoch_verify_tests::build_table_legs
-struct BatchedTableLeg {
-    deep: super::deep::DeepShape,
-    analysis: super::constraints::Analysis,
-    quotient: super::constraints::QuotientShape,
-    main_width: usize,
-    num_alpha_powers: usize,
-}
-
-fn batched_leg_shapes(e: &RealBatchedEpoch) -> Vec<BatchedTableLeg> {
-    use stark::verifier::{IsStarkVerifier, Verifier};
-
-    e.refs()
-        .iter()
-        .zip(&e.proof.tables)
-        .map(|(air, data)| {
-            let layout = Verifier::<Gl, Ext3, ()>::ood_layout(*air);
-            let artifact = stark::constraint_ir::ConstraintArtifact::capture(*air);
-            let (main_width, aux_width) = air.trace_layout();
-            let num_total_cols = main_width + aux_width;
-            let has_aux = air.has_aux_trace();
-            BatchedTableLeg {
-                deep: super::deep::DeepShape {
-                    step_size: layout.step_size(),
-                    num_eval_points: artifact.shape.transition_offsets.len() * layout.step_size(),
-                    num_total_cols,
-                    next_row_cols: layout.next_row_cols().to_vec(),
-                    num_composition_parts: data.composition_poly_parts_ood_evaluation.len(),
-                    log2_trace_length: data.trace_length.trailing_zeros(),
-                },
-                analysis: super::constraints::analyze(&artifact),
-                quotient: super::constraints::QuotientShape {
-                    log2_trace_length: data.trace_length.trailing_zeros(),
-                    num_composition_parts: data.composition_poly_parts_ood_evaluation.len(),
-                    boundary: super::epoch_verify::boundary_terms(has_aux, num_total_cols),
-                },
-                main_width,
-                num_alpha_powers: if has_aux {
-                    artifact.shape.max_bus_elements as usize
-                } else {
-                    0
-                },
-            }
-        })
-        .collect()
-}
-
-/// Hint `count` consecutive words of `arena`, advancing `cursor` — the
-/// walk over the batched opening arena's declared order.
-fn hint_run(
-    b: &mut LfmBuilder,
-    arena: super::instr::ArenaId,
-    cursor: &mut u32,
-    count: usize,
-) -> Vec<super::builder::Cell> {
-    (0..count)
-        .map(|_| {
-            let c = b.hint_word(arena, *cursor);
-            *cursor += 1;
-            c
-        })
-        .collect()
-}
-
-/// Hint `count` digests (two words each) of `arena`, advancing `cursor`.
-fn hint_digests(
-    b: &mut LfmBuilder,
-    arena: super::instr::ArenaId,
-    cursor: &mut u32,
-    count: usize,
-) -> Vec<super::edsl::WrapDigest> {
-    (0..count)
-        .map(|_| {
-            // The stride is the DIGEST's width, not a literal two.
-            let d = super::edsl::hint_digest(b, arena, *cursor);
-            *cursor += super::edsl::digest_words(b);
-            d
-        })
-        .collect()
-}
-
-/// [`batched_epoch_program`] with the OPENING AUTHENTICATION legs hung off
-/// the spine (`with_openings`), and with the deliberately WRONG index
-/// reduction (`wrong_reduction`) — the machine port of
-/// `short_round_low_bit_convention_is_exercised`: keeping the LOW bits of
-/// the shared index instead of the high ones is self-consistent host-side,
-/// and here it must make an honest proof's walk UNPROVABLE, because the
-/// spine's roots were computed over the other convention.
-pub(super) fn batched_epoch_program_with(
-    e: &RealBatchedEpoch,
-    with_openings: bool,
-    wrong_reduction: bool,
-) -> LfmProgram {
-    use super::batched_epoch::{
-        BatchedEpochAbsorbs, BatchedPrepRoot, BatchedTableOod, emit_batched_epoch_challenges,
-    };
-    use super::statement_replay::{EpochStatementVars, absorb_epoch_statement};
-
-    assert!(
-        !wrong_reduction || with_openings,
-        "the reduction control is a property of the walks"
-    );
-
-    let mut b = LfmBuilder::new().with_wrap_hash(super::edsl::WrapHash::production());
-    let shape = batched_shape_of(e);
-
-    // ---- arenas, declaration order = absorb order ----
-    let stmt_halves = 8 + e.statement.public_output_len.div_ceil(4) + 2;
-    let a_stmt = b.declare_arena(stmt_halves as u32);
-    let num_arena_prep = e
-        .prep_sources
-        .iter()
-        .filter(|p| p.is_some_and(PrepSource::is_arena))
-        .count();
-    let a_prep_roots =
-        b.declare_arena(super::proof_arena::words_per_root() as u32 * num_arena_prep as u32);
-    // The carved root's arena sits between the prep roots and main_root —
-    // declaration order is absorb order, and that is its transcript slot.
-    let a_carved_root = shape
-        .carved_main
-        .map(|_| b.declare_arena(super::proof_arena::words_per_root() as u32));
-    let a_main_root = b.declare_arena(super::proof_arena::words_per_root() as u32);
-    let num_reg = crate::tables::register::NUM_REGISTER_ADDRESSES as u32;
-    let a_reg_init = b.declare_arena(num_reg);
-    let a_reg_fini = b.declare_arena(num_reg);
-    let a_pc_start = b.declare_arena(2);
-    let a_aux_root = shape
-        .has_aux
-        .then(|| b.declare_arena(super::proof_arena::words_per_root() as u32));
-    let a_contrib: Vec<Option<super::instr::ArenaId>> = shape
-        .tables
-        .iter()
-        .map(|t| t.has_contribution.then(|| b.declare_arena(1)))
-        .collect();
-    let a_ood: Vec<(
-        super::instr::ArenaId,
-        super::instr::ArenaId,
-        super::instr::ArenaId,
-    )> = shape
-        .tables
-        .iter()
-        .map(|t| {
-            (
-                b.declare_arena((t.ood_current_dims.0 * t.ood_current_dims.1) as u32),
-                b.declare_arena((t.ood_next_dims.0 * t.ood_next_dims.1) as u32),
-                b.declare_arena(t.num_parts as u32),
-            )
-        })
-        .collect();
-    let a_parts_root = b.declare_arena(super::proof_arena::words_per_root() as u32);
-    // The standalone class's terminal polynomials — per table, sized by the
-    // trace-length degree bound, absorbed in round 4 and evaluated by the
-    // standalone terminal checks.
-    let a_standalone: Vec<Option<super::instr::ArenaId>> = (0..shape.tables.len())
-        .map(|t| {
-            shape
-                .fri
-                .plan
-                .standalone
-                .contains(&t)
-                .then(|| b.declare_arena(1u32 << (shape.heights[t] as u32 - shape.log2_blowup)))
-        })
-        .collect();
-    let a_fri_roots = b.declare_arena(
-        super::proof_arena::words_per_root() as u32 * shape.fri.num_committed() as u32,
-    );
-    let a_fri_coeffs = b.declare_arena(shape.fri.num_terminal_coeffs() as u32);
-    let a_nonce = (shape.grinding_factor > 0).then(|| b.declare_arena(1));
-    // The opening arena, LAST and exactly the T1 serializer's size — the
-    // program declares what `batched_opening_arena` fills, in its order.
-    let a_openings = with_openings.then(|| {
-        b.declare_arena(
-            (e.proof.queries.len()
-                * super::epoch_verify_tests::batched_opening_words_per_query(&e.shape))
-                as u32,
-        )
-    });
-    let a_fri_legs = with_openings.then(|| {
-        b.declare_arena(
-            (e.proof.queries.len()
-                * super::epoch_verify_tests::batched_fri_words_per_query(&e.shape, &e.fri_params))
-                as u32,
-        )
-    });
-
-    // ---- the statement ----
-    let stmt: Vec<_> = (0..stmt_halves as u32)
-        .map(|i| b.hint_felt(a_stmt, i))
-        .collect();
-    let out_halves = e.statement.public_output_len.div_ceil(4);
-    let (elf_digest, rest) = stmt.split_at(8);
-    let (public_output, epoch_label) = rest.split_at(out_halves);
-
-    let mut t = TranscriptReplay::new(&[]);
-    absorb_epoch_statement(
-        &mut t,
-        &e.statement,
-        &EpochStatementVars {
-            elf_digest,
-            public_output,
-            epoch_label,
-        },
-    );
-
-    // ---- registers, and the preprocessed roots from their provenances ----
-    let reg_init: Vec<_> = (0..num_reg).map(|r| b.hint_felt(a_reg_init, r)).collect();
-    let reg_fini: Vec<_> = (0..num_reg).map(|r| b.hint_felt(a_reg_fini, r)).collect();
-    for cell in reg_init.iter().chain(&reg_fini) {
-        super::epoch::assert_u32(&mut b, *cell);
-    }
-    let reg_shape = super::programs::RegisterDerivationShape {
-        blowup: e.opts.blowup_factor as usize,
-        coset_offset: e.opts.coset_offset,
-    };
-
-    let mut next_arena_prep = 0usize;
-    let mut decode_cells: Option<RootCells> = None;
-    let prep_cells: Vec<Option<RootCells>> = e
-        .prep_sources
-        .iter()
-        .map(|prep| match prep {
-            None => None,
-            // Interned: the legs compare against these lanes; the ABSORB still
-            // goes through `BatchedPrepRoot::Constant`'s literal bytes (the
-            // splice economy). Both views are the same program text.
-            Some(PrepSource::Constant(c)) => Some(RootCells::constant(&mut b, c)),
-            Some(PrepSource::Register(_)) => {
-                let digest = super::programs::emit_register_commitment(
-                    &mut b, reg_shape, &reg_init, &reg_fini,
-                );
-                Some(RootCells::from_digest(&mut b, digest))
-            }
-            Some(PrepSource::ElfDependent(_)) => {
-                let cells = RootCells::hint(
-                    &mut b,
-                    a_prep_roots,
-                    super::proof_arena::words_per_root() as u32 * next_arena_prep as u32,
-                );
-                next_arena_prep += 1;
-                assert!(
-                    decode_cells.is_none(),
-                    "a continuation epoch has one ELF-dependent preprocessed root"
-                );
-                decode_cells = Some(cells.clone());
-                Some(cells)
-            }
-        })
-        .collect();
-    assert_eq!(next_arena_prep, num_arena_prep);
-
-    let prep_slots: Vec<Option<BatchedPrepRoot<'_>>> = e
-        .prep_sources
-        .iter()
-        .zip(&prep_cells)
-        .map(|(prep, cells)| match (prep, cells) {
-            (None, _) => None,
-            (Some(PrepSource::Constant(c)), _) => Some(BatchedPrepRoot::Constant(c)),
-            (_, Some(cells)) => Some(BatchedPrepRoot::Cells(cells)),
-            _ => unreachable!("a non-constant prep source has cells"),
-        })
-        .collect();
-
-    // ---- the proof-carried cells ----
-    let carved_cells = a_carved_root.map(|id| RootCells::hint(&mut b, id, 0));
-    let main_cells = RootCells::hint(&mut b, a_main_root, 0);
-    let aux_cells = a_aux_root.map(|id| RootCells::hint(&mut b, id, 0));
-    let contribs: Vec<Option<super::builder::Ext>> = a_contrib
-        .iter()
-        .map(|id| id.map(|id| b.hint_word(id, 0).as_ext()))
-        .collect();
-    let ood_cells: Vec<(Vec<_>, Vec<_>, Vec<_>)> = shape
-        .tables
-        .iter()
-        .zip(&a_ood)
-        .map(|(t, (ac, an, ap))| {
-            (
-                (0..(t.ood_current_dims.0 * t.ood_current_dims.1) as u32)
-                    .map(|k| b.hint_word(*ac, k).as_ext())
-                    .collect(),
-                (0..(t.ood_next_dims.0 * t.ood_next_dims.1) as u32)
-                    .map(|k| b.hint_word(*an, k).as_ext())
-                    .collect(),
-                (0..t.num_parts as u32)
-                    .map(|k| b.hint_word(*ap, k).as_ext())
-                    .collect(),
-            )
-        })
-        .collect();
-    let parts_cells = RootCells::hint(&mut b, a_parts_root, 0);
-    let standalone_cells: Vec<Option<Vec<super::builder::Ext>>> = a_standalone
-        .iter()
-        .enumerate()
-        .map(|(t, id)| {
-            id.map(|id| {
-                (0..1u32 << (shape.heights[t] as u32 - shape.log2_blowup))
-                    .map(|k| b.hint_word(id, k).as_ext())
-                    .collect()
-            })
-        })
-        .collect();
-    let fri_root_cells: Vec<_> = (0..shape.fri.num_committed())
-        .map(|k| {
-            RootCells::hint(
-                &mut b,
-                a_fri_roots,
-                super::proof_arena::words_per_root() as u32 * k as u32,
-            )
-        })
-        .collect();
-    let coeff_cells: Vec<_> = (0..shape.fri.num_terminal_coeffs() as u32)
-        .map(|k| b.hint_word(a_fri_coeffs, k).as_ext())
-        .collect();
-    let nonce = a_nonce.map(|id| b.hint_felt(id, 0));
-
-    // ---- the ONE-transcript spine ----
-    let oods: Vec<BatchedTableOod<'_>> = ood_cells
-        .iter()
-        .map(|(c, x, p)| BatchedTableOod {
-            current: c,
-            next: x,
-            parts: p,
-        })
-        .collect();
-    let ch = emit_batched_epoch_challenges(
-        &mut b,
-        &mut t,
-        &shape,
-        &BatchedEpochAbsorbs {
-            prep_roots: &prep_slots,
-            carved_root: carved_cells.as_ref(),
-            main_root: &main_cells,
-            aux_root: aux_cells.as_ref(),
-            contributions: &contribs,
-            parts_root: &parts_cells,
-            ood: &oods,
-            standalone_coeffs: &standalone_cells,
-            fri_roots: &fri_root_cells,
-            fri_coeffs: &coeff_cells,
-            nonce,
-        },
-    );
-
-    // ---- publishes: the pair, then the attestation, then every challenge ----
-    b.public(ch.lookup.0.as_cell());
-    b.public(ch.lookup.1.as_cell());
-    {
-        let pc_start: Vec<_> = (0..2).map(|i| b.hint_felt(a_pc_start, i)).collect();
-        let decode = decode_cells
-            .as_ref()
-            .expect("a continuation epoch has a DECODE sub-proof")
-            .byte_halves(&mut b);
-        let id = super::programs::emit_program_id(
-            &mut b,
-            super::programs::ProgramIdShape { num_pages: 0 },
-            elf_digest,
-            &pc_start,
-            &decode,
-            &[],
-        );
-        b.public(id[0]);
-        b.public(id[1]);
-    }
-    for v in ch.betas.iter().chain(&ch.zs).chain(&ch.gammas) {
-        b.public(v.as_cell());
-    }
-    b.public(ch.alpha.as_cell());
-    for zeta in &ch.zetas {
-        b.public(zeta.as_cell());
-    }
-    for bits in &ch.iota_bits {
-        let felt = edsl::bits_to_felt(&mut b, bits);
-        b.public(felt.as_cell());
-    }
-
-    // ---- the LogUp closure, on the cells the spine absorbed ----
-    let contributions: Vec<super::builder::Ext> = contribs.iter().copied().flatten().collect();
-    let lshape = super::logup::LogUpShape {
-        num_contributing_tables: contributions.len(),
-        num_output_bytes: e.statement.public_output_len,
-    };
-    let start = reg_init[crate::tables::register::X254_INDEX];
-    let bytes = super::epoch::emit_output_bytes(&mut b, public_output, lshape.num_output_bytes);
-    let target = super::logup::emit_commit_bus_target(
-        &mut b,
-        &lshape,
-        ch.lookup.0,
-        ch.lookup.1,
-        start,
-        &bytes,
-    );
-    let total = super::logup::emit_bus_closure(&mut b, &lshape, &contributions, target);
-    b.public(total.as_cell());
-
-    // ---- the aggregator-facing published-word schema (carved programs only
-    // — the continuation batched format). P3's aggregator byte-compares
-    // these across the five wraps and against the global proof; publishing
-    // them here is what makes cross-epoch chaining a check on PUBLISHED
-    // words instead of a trust. Order of record: register boundary vectors
-    // (init then fini), the epoch label, the epoch's output bytes, then the
-    // carved L2G root. Every cell is already program state (hinted once,
-    // bound by the walks/statement above) — publishing adds no arena words
-    // and no wrap-hash permutations.
-    if shape.carved_main.is_some() {
-        for cell in reg_init.iter().chain(&reg_fini) {
-            b.public(cell.as_cell());
-        }
-        for half in epoch_label {
-            b.public(half.as_cell());
-        }
-        for byte in &bytes {
-            b.public(byte.as_cell());
-        }
-        let carved = carved_cells
-            .as_ref()
-            .expect("a carved epoch has carved root cells");
-        for half in carved.lanes_flat() {
-            b.public(half.as_cell());
-        }
-    }
-
-    // ---- the opening walks: every round authenticated at the REDUCED shared
-    // index, against the very root cells the spine absorbed ----
-    if let Some(a_open) = a_openings {
-        use super::batched_epoch_verify::{
-            MixedMatrixOpening, emit_mixed_verify_batch, reduce_iota_bits,
-        };
-        use super::sub_proof::{GroupCommitment, GroupOpening, GroupShape};
-
-        use super::deep::DeepOpening;
-
-        let n = e.proof.tables.len();
-        let h_max_fri = e.shape.heights.iter().copied().max().expect("tables");
-
-        // Every table's matrix position in each round — the crossing's read,
-        // the same one `table_deep_pairs` makes.
-        let prep_pos: Vec<Option<usize>> = (0..n)
-            .map(|t| e.shape.prep.tables.iter().position(|&x| x == t))
-            .collect();
-        // `None` exactly for a carved table, whose main row pair comes from
-        // its standalone walk instead of the mixed main round.
-        let main_pos: Vec<Option<usize>> = (0..n)
-            .map(|t| e.shape.main.tables.iter().position(|&x| x == t))
-            .collect();
-        let aux_pos: Vec<Option<usize>> = (0..n)
-            .map(|t| e.shape.aux.tables.iter().position(|&x| x == t))
-            .collect();
-        let parts_pos: Vec<usize> = (0..n)
-            .map(|t| {
-                e.shape
-                    .parts
-                    .tables
-                    .iter()
-                    .position(|&x| x == t)
-                    .expect("every table has a parts matrix")
-            })
-            .collect();
-
-        // Per table, hoisted across queries: the OOD grid rebuilt from the
-        // very cells the spine absorbed (no second copy for a prover to
-        // disagree with), the CONSTRAINT identity and the quotient check at
-        // this table's z — production's own evaluators, reached through
-        // emit_analyzed/emit_quotient exactly as the per-table program does —
-        // and the DEEP invariants over the same grid and the same parts.
-        let legs = batched_leg_shapes(e);
-        let dinvs: Vec<super::deep::DeepInvariants> = (0..n)
-            .map(|t_i| {
-                let leg = &legs[t_i];
-                let grid = super::epoch::emit_reconstruct_ood(
-                    &mut b,
-                    &leg.deep,
-                    &ood_cells[t_i].0,
-                    &ood_cells[t_i].1,
-                );
-
-                // The LogUp uniforms, DERIVED: α powers from the one shared α
-                // the spine sampled, and the per-row offset from the one `L`
-                // it absorbed — the same cell the closure sums.
-                let alpha_powers = if leg.num_alpha_powers > 0 {
-                    super::constraints::emit_alpha_powers(&mut b, ch.lookup.1, leg.num_alpha_powers)
-                } else {
-                    Vec::new()
-                };
-                let table_offset = match contribs[t_i] {
-                    Some(l) => super::constraints::emit_table_offset(
-                        &mut b,
-                        l,
-                        leg.quotient.log2_trace_length,
-                    ),
-                    None => b.felt_const(FE::zero()).as_ext(),
-                };
-                let steps = super::epoch_verify::frame_step_view(&grid, leg.deep.step_size);
-                let ood_ops = super::constraints::OodOperands {
-                    steps,
-                    main_width: leg.main_width,
-                    rap_challenges: vec![ch.lookup.0, ch.lookup.1],
-                    alpha_powers,
-                    table_offset,
-                };
-                let evals = super::constraints::emit_analyzed(&mut b, &leg.analysis, &ood_ops);
-                let q = super::constraints::emit_quotient(
-                    &mut b,
-                    &leg.quotient,
-                    &ood_ops,
-                    ch.zs[t_i],
-                    ch.betas[t_i],
-                    &evals,
-                    &ood_cells[t_i].2,
-                );
-                b.assert_eq_ext(q.claimed, q.composition);
-
-                super::deep::emit_deep_invariants(
-                    &mut b,
-                    &leg.deep,
-                    ch.gammas[t_i],
-                    ch.zs[t_i],
-                    &grid,
-                    &ood_cells[t_i].2,
-                )
-            })
-            .collect();
-        let fri_layer_commitments: Vec<super::fri::LayerCommitment> = fri_root_cells
-            .iter()
-            .map(|c| super::fri::LayerCommitment {
-                root_lanes: c.lanes.clone(),
-            })
-            .collect();
-
-        let mut cursor: u32 = 0;
-        let mut fri_cursor: u32 = 0;
-        for bits in &ch.iota_bits {
-            // ---- the walks: preprocessed tables, then the mixed rounds ----
-            let mut prep_values: Vec<Vec<super::builder::Cell>> = Vec::new();
-            for (slot, &(h, w)) in e.shape.prep.tables.iter().zip(e.shape.prep.dims.iter()) {
-                let cells = prep_cells[*slot]
-                    .as_ref()
-                    .expect("a preprocessed table has root cells");
-                let values = hint_run(&mut b, a_open, &mut cursor, 2 * w);
-                let siblings = hint_digests(&mut b, a_open, &mut cursor, h - 1);
-                let tbits = if wrong_reduction && h < h_max_fri {
-                    // ★ THE BROKEN CONTROL: keep the LOW bits instead of the
-                    // high ones — same length, wrong index space.
-                    &bits[..h - 1]
-                } else {
-                    reduce_iota_bits(bits, h_max_fri, h)
-                };
-                super::sub_proof::emit_group_authentication(
-                    &mut b,
-                    &GroupCommitment::from_lanes(
-                        cells.lanes.clone(),
-                        GroupShape {
-                            num_columns: w,
-                            is_ext: false,
-                        },
-                    ),
-                    &GroupOpening {
-                        values: values.clone(),
-                        siblings,
-                    },
-                    tbits,
-                );
-                prep_values.push(values);
-            }
-
-            // ---- the carved table's standalone walk: the preprocessed
-            // pattern with the root PROOF-CARRIED — authenticated against the
-            // very cells the spine absorbed, at the reduced shared index. The
-            // wrong-reduction control covers this walk exactly as it covers
-            // the others (verdict condition 4's emitted side).
-            let mut carved_values: Option<Vec<super::builder::Cell>> = None;
-            if let Some((ct, cw)) = shape.carved_main {
-                let cells = carved_cells
-                    .as_ref()
-                    .expect("a carved epoch has carved root cells");
-                let h = e.shape.heights[ct];
-                let values = hint_run(&mut b, a_open, &mut cursor, 2 * cw);
-                let siblings = hint_digests(&mut b, a_open, &mut cursor, h - 1);
-                let tbits = if wrong_reduction && h < h_max_fri {
-                    &bits[..h - 1]
-                } else {
-                    reduce_iota_bits(bits, h_max_fri, h)
-                };
-                super::sub_proof::emit_group_authentication(
-                    &mut b,
-                    &GroupCommitment::from_lanes(
-                        cells.lanes.clone(),
-                        GroupShape {
-                            num_columns: cw,
-                            is_ext: false,
-                        },
-                    ),
-                    &GroupOpening {
-                        values: values.clone(),
-                        siblings,
-                    },
-                    tbits,
-                );
-                carved_values = Some(values);
-            }
-
-            // The three mixed rounds: per-matrix row pairs in round INPUT
-            // order, then the round's ONE shared path. The value cells are
-            // KEPT — the crossing below reads the cells the walks
-            // authenticated, never a second copy.
-            let mut round_values: Vec<Vec<Vec<super::builder::Cell>>> = Vec::new();
-            let mut rounds: Vec<(&stark::batched::shape::RoundShape, &RootCells, bool)> =
-                vec![(&e.shape.main, &main_cells, false)];
-            if let Some(aux) = aux_cells.as_ref() {
-                rounds.push((&e.shape.aux, aux, true));
-            }
-            rounds.push((&e.shape.parts, &parts_cells, true));
-            for (round, root, is_ext) in rounds {
-                let h_round = round.h_max().expect("a committed round is non-empty");
-                let per_values: Vec<Vec<super::builder::Cell>> = round
-                    .dims
-                    .iter()
-                    .map(|&(_, w)| hint_run(&mut b, a_open, &mut cursor, 2 * w))
-                    .collect();
-                let siblings = hint_digests(&mut b, a_open, &mut cursor, h_round - 1);
-                let matrices: Vec<MixedMatrixOpening<'_>> = round
-                    .dims
-                    .iter()
-                    .zip(&per_values)
-                    .map(|(&(h, w), values)| MixedMatrixOpening {
-                        shape: GroupShape {
-                            num_columns: w,
-                            is_ext,
-                        },
-                        log_height: h,
-                        values,
-                    })
-                    .collect();
-                let rbits = if wrong_reduction && h_round < h_max_fri {
-                    &bits[..h_round - 1]
-                } else {
-                    reduce_iota_bits(bits, h_max_fri, h_round)
-                };
-                emit_mixed_verify_batch(&mut b, root, &matrices, &siblings, rbits);
-                round_values.push(per_values);
-            }
-            let main_values = &round_values[0];
-            let aux_values = aux_cells.as_ref().map(|_| &round_values[1]);
-            let parts_values = round_values.last().expect("the parts round");
-
-            // ---- the crossing: per table, the authenticated cells re-read
-            // by POINT, folded to the DEEP pair at the reduced index ----
-            let mut points: Vec<(super::builder::Felt, super::builder::Felt)> =
-                Vec::with_capacity(n);
-            let mut deep_pairs: Vec<(super::builder::Ext, super::builder::Ext)> =
-                Vec::with_capacity(n);
-            for t_i in 0..n {
-                let h_t = e.shape.heights[t_i];
-                let rbits = reduce_iota_bits(bits, h_max_fri, h_t);
-                let (point, point_sym) = super::sub_proof::emit_points_from_bits(
-                    &mut b,
-                    h_t as u32,
-                    shape.coset_offset,
-                    rbits,
-                );
-
-                let mut trace = Vec::with_capacity(legs[t_i].deep.num_total_cols);
-                let mut trace_sym = Vec::with_capacity(legs[t_i].deep.num_total_cols);
-                if let Some(m) = prep_pos[t_i] {
-                    let w = e.shape.prep.dims[m].1;
-                    let vals = &prep_values[m];
-                    trace.extend((0..w).map(|c| vals[c].as_ext()));
-                    trace_sym.extend((0..w).map(|c| vals[w + c].as_ext()));
-                }
-                match main_pos[t_i] {
-                    Some(m) => {
-                        let w = e.shape.main.dims[m].1;
-                        let vals = &main_values[m];
-                        trace.extend((0..w).map(|c| vals[c].as_ext()));
-                        trace_sym.extend((0..w).map(|c| vals[w + c].as_ext()));
-                    }
-                    None => {
-                        let (ct, cw) = shape
-                            .carved_main
-                            .expect("only a carved table lacks a main matrix");
-                        assert_eq!(ct, t_i, "the carved table is the one without a main slot");
-                        let vals = carved_values
-                            .as_ref()
-                            .expect("the carved walk authenticated this query");
-                        trace.extend((0..cw).map(|c| vals[c].as_ext()));
-                        trace_sym.extend((0..cw).map(|c| vals[cw + c].as_ext()));
-                    }
-                }
-                if let Some(m) = aux_pos[t_i] {
-                    let w = e.shape.aux.dims[m].1;
-                    let vals = &aux_values.expect("an aux position implies an aux round")[m];
-                    trace.extend((0..w).map(|c| vals[c].as_ext()));
-                    trace_sym.extend((0..w).map(|c| vals[w + c].as_ext()));
-                }
-                assert_eq!(
-                    trace.len(),
-                    legs[t_i].deep.num_total_cols,
-                    "the crossing must cover exactly the DEEP column set"
-                );
-                let m = parts_pos[t_i];
-                let w = e.shape.parts.dims[m].1;
-                assert_eq!(
-                    w, legs[t_i].deep.num_composition_parts,
-                    "the parts matrix is one column per composition part"
-                );
-                let vals = &parts_values[m];
-                let parts: Vec<super::builder::Ext> = (0..w).map(|c| vals[c].as_ext()).collect();
-                let parts_sym: Vec<super::builder::Ext> =
-                    (0..w).map(|c| vals[w + c].as_ext()).collect();
-
-                let regular = DeepOpening {
-                    point,
-                    trace,
-                    parts,
-                };
-                let symmetric = DeepOpening {
-                    point: point_sym,
-                    trace: trace_sym,
-                    parts: parts_sym,
-                };
-                deep_pairs.push((
-                    super::deep::emit_deep_point(
-                        &mut b,
-                        &legs[t_i].deep,
-                        ch.gammas[t_i],
-                        &dinvs[t_i],
-                        &regular,
-                    ),
-                    super::deep::emit_deep_point(
-                        &mut b,
-                        &legs[t_i].deep,
-                        ch.gammas[t_i],
-                        &dinvs[t_i],
-                        &symmetric,
-                    ),
-                ));
-                points.push((point, point_sym));
-            }
-
-            // ---- the mix, the batched instance, the standalone class ----
-            let (p0, p0_sym, buckets) = super::batched_epoch_verify::emit_query_mix(
-                &mut b,
-                &shape.fri.plan.batched,
-                &e.shape.heights,
-                h_max_fri,
-                ch.alpha,
-                &deep_pairs,
-                bits,
-            );
-            // υ in the TALLEST domain is the tallest table's own point — its
-            // reduction is the identity, so reusing the cell adds no second
-            // derivation.
-            let tallest = e
-                .shape
-                .heights
-                .iter()
-                .position(|&h| h == h_max_fri)
-                .expect("a tallest table exists");
-            let fri_openings_q: Vec<super::fri::LayerOpening> = (0..shape.fri.num_committed())
-                .map(|i| {
-                    let sym = {
-                        let c = b.hint_word(
-                            a_fri_legs.expect("the FRI arena exists with the legs"),
-                            fri_cursor,
-                        );
-                        fri_cursor += 1;
-                        c.as_ext()
-                    };
-                    let siblings = hint_digests(
-                        &mut b,
-                        a_fri_legs.expect("the FRI arena exists with the legs"),
-                        &mut fri_cursor,
-                        h_max_fri - i - 2,
-                    );
-                    super::fri::LayerOpening { sym, siblings }
-                })
-                .collect();
-            super::batched_epoch_verify::emit_batched_query_fri(
-                &mut b,
-                &shape.fri.layout,
-                h_max_fri,
-                &fri_layer_commitments,
-                &ch.zetas,
-                &coeff_cells,
-                bits,
-                points[tallest].0,
-                points[tallest].1,
-                p0,
-                p0_sym,
-                &buckets,
-                &fri_openings_q,
-            );
-            for &t_i in &shape.fri.plan.standalone {
-                let coeffs = standalone_cells[t_i]
-                    .as_ref()
-                    .expect("a standalone table has terminal cells");
-                super::batched_epoch_verify::emit_standalone_terminal_check(
-                    &mut b,
-                    coeffs,
-                    points[t_i].0,
-                    points[t_i].1,
-                    deep_pairs[t_i].0,
-                    deep_pairs[t_i].1,
-                );
-            }
-        }
-        assert_eq!(
-            cursor as usize,
-            e.proof.queries.len()
-                * super::epoch_verify_tests::batched_opening_words_per_query(&e.shape),
-            "the walks must consume exactly the declared opening arena"
-        );
-        assert_eq!(
-            fri_cursor as usize,
-            e.proof.queries.len()
-                * super::epoch_verify_tests::batched_fri_words_per_query(&e.shape, &e.fri_params),
-            "the FRI legs must consume exactly the declared arena"
-        );
-    }
-
-    let program = compile(b.finish());
-    validate(&program).expect("the batched epoch spine must be admissible");
-    program
-}
-
-/// The arenas [`batched_epoch_program`] declares, in the same order, filled
-/// from the harness's proof.
-pub(super) fn batched_epoch_arenas(e: &RealBatchedEpoch) -> Vec<Vec<LfmWord>> {
-    let mut stmt: Vec<FE> = Vec::new();
-    let halves = |bytes: &[u8]| -> Vec<FE> {
-        bytes
-            .chunks(4)
-            .map(|c| {
-                let mut w = [0u8; 4];
-                w[..c.len()].copy_from_slice(c);
-                FE::from(u32::from_le_bytes(w) as u64)
-            })
-            .collect()
-    };
-    stmt.extend(halves(&e.elf_digest));
-    stmt.extend(halves(&e.public_output));
-    stmt.extend(halves(&e.epoch_label.to_le_bytes()));
-
-    let prep: Vec<Commitment> = e
-        .prep_sources
-        .iter()
-        .filter_map(|p| match p {
-            Some(PrepSource::ElfDependent(c)) => Some(*c),
-            _ => None,
-        })
-        .collect();
-    let reg = |v: &[u32]| -> Vec<LfmWord> {
-        assert_eq!(
-            v.len(),
-            crate::tables::register::NUM_REGISTER_ADDRESSES,
-            "a register boundary vector is one word per register word address"
-        );
-        v.iter()
-            .map(|w| base_word(FE::from(u64::from(*w))))
-            .collect()
-    };
-
-    let mut out = vec![
-        stmt.iter().map(|h| base_word(*h)).collect(),
-        super::proof_arena::commitments_to_arena(&prep),
-    ];
-    // The carved root's arena sits between the prep roots and main_root —
-    // the program's declaration order is the absorb order.
-    if e.shape.carved_main.is_some() {
-        out.push(super::proof_arena::commitments_to_arena(&[e
-            .proof
-            .carved_main_root
-            .expect("a carved epoch proof carries its carved root")]));
-    }
-    out.extend([
-        super::proof_arena::commitments_to_arena(&[e.proof.main_root]),
-        reg(&e.register_init),
-        reg(&e.reg_fini),
-        super::keccak_host::pack_stream(&e.pc_start.to_le_bytes())
-            .into_iter()
-            .map(base_word)
-            .collect(),
-    ]);
-    if let Some(root) = e.proof.aux_root.as_ref() {
-        out.push(super::proof_arena::commitments_to_arena(&[*root]));
-    }
-    for table in &e.proof.tables {
-        if let Some(bpi) = table.bus_public_inputs.as_ref() {
-            out.push(vec![ext_word(&bpi.table_contribution)]);
-        }
-    }
-    for table in &e.proof.tables {
-        let block_words = |block: &stark::table::Table<Ext3>| -> Vec<LfmWord> {
-            (0..block.height)
-                .flat_map(|r| block.get_row(r).to_vec())
-                .map(|v| ext_word(&v))
-                .collect()
-        };
-        out.push(block_words(&table.trace_ood_evaluations));
-        out.push(block_words(&table.trace_ood_next_evaluations));
-        out.push(
-            table
-                .composition_poly_parts_ood_evaluation
-                .iter()
-                .map(ext_word)
-                .collect(),
-        );
-    }
-    out.push(super::proof_arena::commitments_to_arena(&[e
-        .proof
-        .parts_root]));
-    for table in &e.proof.tables {
-        if let Some(coeffs) = table.standalone_final_poly_coeffs.as_ref() {
-            out.push(coeffs.iter().map(ext_word).collect());
-        }
-    }
-    out.push(super::proof_arena::commitments_to_arena(
-        &e.proof.fri_layer_roots,
-    ));
-    out.push(e.proof.fri_final_poly_coeffs.iter().map(ext_word).collect());
-    if let Some(nc) = e.proof.nonce {
-        out.push(vec![base_word(FE::from(nc))]);
-    }
-    out
-}
-
-/// ★ THE RUN: the BATCHED epoch's Fiat-Shamir spine, executed against a real
-/// batched epoch proof the host verification accepts, and differentialled
-/// against `replay_epoch_transcript`'s own challenges — every β, z, γ, the
-/// shared pair, the shared α, every ζ, every shared iota, the attestation
-/// program_id, and the COMMIT-bus closure.
-#[test]
-fn the_batched_epoch_challenge_spine_matches_production() {
-    let e = real_batched_epoch_with(super::proof_fixture::fixture_options());
-    let program = batched_epoch_program(&e);
-    let arenas = batched_epoch_arenas(&e);
-    let exec = execute(&program, &arenas, &crate::hash_pin::BLOCK_HASHER)
-        .expect("the batched epoch spine must execute");
-
-    // Vacuity guard: the fixture must exercise BOTH instance classes, or the
-    // standalone-terminal absorb and the class split are dead paths here.
-    assert!(
-        !e.challenges.fri.plan.standalone.is_empty() && !e.challenges.fri.plan.batched.is_empty(),
-        "the fixture epoch must have both batched and standalone tables"
-    );
-
-    let pub_ext = |i: usize| word_as_ext(&exec.public_words[i].1).expect("an ext challenge");
-    let [z, alpha] = e.challenges.lookup.as_slice() else {
-        panic!("the shared pair is (z, α)");
-    };
-    assert_eq!(pub_ext(0), *z, "the shared LogUp z");
-    assert_eq!(pub_ext(1), *alpha, "the shared LogUp alpha");
-    assert_eq!(
-        published_digest(&exec.public_words, 2),
-        e.expected_program_id,
-        "the attestation program_id must match production's"
-    );
-
-    let n = e.proof.tables.len();
-    let mut cursor = 4usize;
-    for (i, want) in e.challenges.betas.iter().enumerate() {
-        assert_eq!(pub_ext(cursor + i), *want, "beta of table {i}");
-    }
-    cursor += n;
-    for (i, want) in e.challenges.zs.iter().enumerate() {
-        assert_eq!(pub_ext(cursor + i), *want, "z of table {i}");
-    }
-    cursor += n;
-    for (i, want) in e.challenges.deep_gammas.iter().enumerate() {
-        assert_eq!(pub_ext(cursor + i), *want, "gamma of table {i}");
-    }
-    cursor += n;
-    assert_eq!(pub_ext(cursor), e.challenges.fri.alpha, "the shared DEEP α");
-    cursor += 1;
-    for (k, want) in e.challenges.fri.betas.iter().enumerate() {
-        assert_eq!(pub_ext(cursor + k), *want, "zeta {k}");
-    }
-    cursor += e.challenges.fri.betas.len();
-    for (q, want) in e.challenges.fri.iotas.iter().enumerate() {
-        let w = exec.public_words[cursor + q].1;
-        let got = super::word::word_as_base(&w).expect("an index is a base felt");
-        assert_eq!(got, FE::from(*want as u64), "shared iota {q}");
-    }
-    cursor += e.challenges.fri.iotas.len();
-    assert_eq!(
-        word_as_ext(&exec.public_words[cursor].1).expect("the bus total is ext"),
-        e.expected_bus_balance,
-        "the LogUp closure must reach production's COMMIT-bus target"
-    );
-    cursor += 1;
-    assert_eq!(
-        cursor,
-        exec.public_words.len(),
-        "every published word must be checked"
-    );
-}
-
-/// ★ THE RUN: the whole ASSEMBLED BATCHED epoch verifier — spine and legs —
-/// on a real continuation epoch proved through the batched path.
-///
-/// What executing proves, stated precisely. Every check is an assert inside
-/// the program, so reaching the end means: all 25 constraint identities and
-/// quotient checks held at the batched spine's own z and β; every round's
-/// opened row pairs hashed — tallest matrices batched, shorter height groups
-/// INJECTED — into the roots the ONE transcript absorbed, at the reduced
-/// shared index; every preprocessed table authenticated against the AIR-set
-/// root its provenance admits; the per-table DEEP crossings fed the α-mixed
-/// injected FRI fold to the batched terminal; every standalone table's
-/// transcript-BOUND polynomial matched its DEEP pair at its own reduced
-/// points; and the LogUp closure reached production's COMMIT-bus target. The
-/// tamper arms and the wrong-reduction control show what does NOT execute,
-/// which is what turns "it executed" into evidence.
-#[test]
-fn the_assembled_batched_epoch_verifier_runs() {
-    let e = real_batched_epoch_with(super::proof_fixture::fixture_options());
-    let program = batched_epoch_program_with(&e, true, false);
-    let mut arenas = batched_epoch_arenas(&e);
-    arenas.push(super::epoch_verify_tests::batched_opening_arena(&e));
-    arenas.push(super::epoch_verify_tests::batched_fri_arena(&e));
-    execute(&program, &arenas, &crate::hash_pin::BLOCK_HASHER)
-        .expect("every opening of an honest batched epoch must authenticate");
-
-    // A moved opening VALUE is unprovable (the first arena word is the first
-    // preprocessed table's first evaluation) — and since the crossing folds
-    // the SAME cell, a value that somehow re-authenticated would still move
-    // the DEEP pair and die at the FRI terminal.
-    let open_idx = arenas.len() - 2;
-    let fri_idx = arenas.len() - 1;
-    let mut tampered = arenas.clone();
-    tampered[open_idx][0] = base_word(FE::from(999_999u64));
-    assert!(
-        execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-        "a tampered opening value must not authenticate"
-    );
-
-    // A moved SIBLING is unprovable (the last arena word is path data — every
-    // round ends with its shared path).
-    let mut tampered = arenas.clone();
-    let last = tampered[open_idx].len() - 1;
-    tampered[open_idx][last] = base_word(FE::from(999_999u64));
-    assert!(
-        execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-        "a tampered sibling must not authenticate"
-    );
-
-    // A moved value in an INJECTED matrix — a main-round matrix SHORTER than
-    // the round's tallest, so both the injection compress in the walk and the
-    // α-mix bucket in the FRI join read it.
-    {
-        let mut off = 0usize;
-        for &(h, w) in &e.shape.prep.dims {
-            off += 2 * w + 2 * (h - 1);
-        }
-        let h_main = e.shape.main.h_max().expect("the main round is non-empty");
-        match e.shape.main.dims.iter().position(|&(h, _)| h < h_main) {
-            Some(m) => {
-                for &(_, w) in &e.shape.main.dims[..m] {
-                    off += 2 * w;
-                }
-                let mut tampered = arenas.clone();
-                tampered[open_idx][off] = base_word(FE::from(999_999u64));
-                assert!(
-                    execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-                    "a tampered injected-matrix value must not verify"
-                );
-            }
-            None => eprintln!("injected-matrix arm skipped: all main matrices are tallest"),
-        }
-    }
-
-    // A moved FRI layer value (the first FRI arena word is layer 0's
-    // symmetric evaluation) must fail its layer walk — or, had it somehow
-    // re-authenticated, the fold chain's terminal.
-    if !arenas[fri_idx].is_empty() {
-        let mut tampered = arenas.clone();
-        tampered[fri_idx][0] = base_word(FE::from(999_999u64));
-        assert!(
-            execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-            "a tampered FRI layer opening must not verify"
-        );
-    }
-
-    // A moved STANDALONE terminal coefficient shifts the transcript (it is
-    // absorbed — the binding the campaign's soundness fix added) AND the
-    // polynomial the standalone check evaluates; both directions kill it.
-    if !e.challenges.fri.plan.standalone.is_empty() {
-        // Position: statement, prep, main_root, reg_init, reg_fini, pc_start,
-        // [aux_root], per-RAP-table contribution, per-table (ood_c, ood_n,
-        // parts), parts_root, THEN the standalone arenas. Count forward.
-        let n = e.proof.tables.len();
-        let num_contrib = e
-            .proof
-            .tables
-            .iter()
-            .filter(|t| t.bus_public_inputs.is_some())
-            .count();
-        let idx = 6 + usize::from(e.proof.aux_root.is_some()) + num_contrib + 3 * n + 1;
-        // The arm must point at what it claims to tamper — pinned by size,
-        // since tampering ANY absorbed arena also fails and would mask a
-        // wrong index.
-        let first_standalone = e.challenges.fri.plan.standalone[0];
-        assert_eq!(
-            arenas[idx].len(),
-            1usize << (e.shape.heights[first_standalone] as u32 - e.fri_params.blowup_log),
-            "the tamper arm must point at the first standalone terminal arena"
-        );
-        let mut tampered = arenas.clone();
-        tampered[idx][0] = ext_word(&FEE::from(999_999u64));
-        assert!(
-            execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-            "a tampered standalone terminal coefficient must not verify"
-        );
-    }
-
-    // ★ The index-reduction DIRECTION (`fri/mmcs.rs`'s convention section):
-    // keeping the low bits instead of the high ones is self-consistent
-    // host-side, so nothing there rejects it; against real roots the walk
-    // must be unprovable. Discrimination is checked, not hoped for: the arm
-    // only proves something when some short walk's two slices actually
-    // differ for this proof's drawn indices.
-    let h_max_fri = e.shape.heights.iter().copied().max().expect("tables");
-    let discriminates = e.challenges.fri.iotas.iter().any(|&iota| {
-        e.shape.prep.dims.iter().any(|&(h, _)| {
-            h < h_max_fri && (iota >> (h_max_fri - h)) != (iota & ((1usize << (h - 1)) - 1))
-        })
-    });
-    if discriminates {
-        let wrong = batched_epoch_program_with(&e, true, true);
-        assert!(
-            execute(&wrong, &arenas, &crate::hash_pin::BLOCK_HASHER).is_err(),
-            "the wrong reduction direction must not authenticate an honest epoch"
-        );
-    } else {
-        // Astronomically unlikely (every short walk's high and low slices
-        // coincide at every query), but a vacuous control must say so rather
-        // than pass silently.
-        eprintln!("wrong-reduction control skipped: the drawn indices do not discriminate");
-    }
-}
-
-/// Arena words the BATCHED epoch program MUST declare, as arithmetic over the
-/// epoch's shapes — `expected_arena_words`' discipline on the batched schema.
-/// Every term comes from the harness's host data (the proof the host
-/// verification accepted, the replayed shape and params), never from the
-/// emitter, so the comparison against the compiled program is absolute.
-fn expected_batched_arena_words(e: &RealBatchedEpoch, with_legs: bool) -> usize {
-    let num_reg = crate::tables::register::NUM_REGISTER_ADDRESSES;
-    // A root's width in arena words (see `expected_arena_words`).
-    let dw = super::proof_arena::words_per_root();
-    let mut total = 8 + e.statement.public_output_len.div_ceil(4) + 2;
-    total += dw
-        * e.prep_sources
-            .iter()
-            .filter(|p| p.is_some_and(PrepSource::is_arena))
-            .count();
-    total += dw * usize::from(e.shape.carved_main.is_some()); // the carved root
-    total += dw; // main_root — ONE, which is the whole batched economy
-    total += 2 * num_reg;
-    total += 2; // pc_start
-    total += dw * usize::from(e.proof.aux_root.is_some());
-    total += e
-        .proof
-        .tables
-        .iter()
-        .filter(|t| t.bus_public_inputs.is_some())
-        .count();
-    for t in &e.proof.tables {
-        total += t.trace_ood_evaluations.width * t.trace_ood_evaluations.height;
-        total += t.trace_ood_next_evaluations.width * t.trace_ood_next_evaluations.height;
-        total += t.composition_poly_parts_ood_evaluation.len();
-    }
-    total += dw; // parts_root
-    for t in &e.proof.tables {
-        if let Some(coeffs) = t.standalone_final_poly_coeffs.as_ref() {
-            total += coeffs.len();
-        }
-    }
-    total += dw * e.proof.fri_layer_roots.len();
-    total += e.proof.fri_final_poly_coeffs.len();
-    total += usize::from(e.fri_params.grinding_factor > 0);
-    if with_legs {
-        total += e.proof.queries.len()
-            * (super::epoch_verify_tests::batched_opening_words_per_query(&e.shape)
-                + super::epoch_verify_tests::batched_fri_words_per_query(&e.shape, &e.fri_params));
-    }
-    total
-}
-
-/// ★ The two ABSOLUTE structural guards, on the BATCHED program — the same
-/// pair that closes the two-consumer class for the per-table one
-/// ([`the_spine_hints_each_proof_value_once`] and
-/// [`the_assembled_verifier_declares_exactly_the_shape_words`]): no arena
-/// word is read twice, every declared word is read, and the schema is
-/// exactly the epoch's shapes — a surplus word is where a second copy of a
-/// joined value (a root, a contribution, a standalone terminal) would hide.
-#[test]
-fn the_batched_verifier_declares_and_hints_exactly_the_shape_words() {
-    use std::collections::HashMap;
-
-    let e = real_batched_epoch_with(super::proof_fixture::fixture_options());
-    for with_legs in [false, true] {
-        let program = batched_epoch_program_with(&e, with_legs, false);
-        let declared: usize = program.arena_schema.lens.iter().map(|l| *l as usize).sum();
-        assert_eq!(
-            declared,
-            expected_batched_arena_words(&e, with_legs),
-            "with_legs = {with_legs}: the batched arena schema must be exactly \
-             the epoch's shapes and nothing more"
-        );
-
-        let mut hints: HashMap<(super::instr::ArenaId, u32), usize> = HashMap::new();
-        for instr in &program.instrs {
-            if let super::instr::Instr::Hint { arena, index, .. } = instr {
-                *hints.entry((*arena, *index)).or_default() += 1;
-            }
-        }
-        let doubled: Vec<_> = hints.iter().filter(|(_, n)| **n > 1).collect();
-        assert!(
-            doubled.is_empty(),
-            "with_legs = {with_legs}: these arena words are hinted more than \
-             once, which is the two-consumer hazard: {doubled:?}"
-        );
-        assert_eq!(
-            hints.len(),
-            declared,
-            "with_legs = {with_legs}: every declared arena word must be read \
-             exactly once"
-        );
-    }
-}
-
-/// ★ The batched query CENSUS: the emitted legs hash exactly what the shape
-/// closed form declares — `batched_query_permutations_for`, checked as the
-/// delta between the with-legs and spine-only programs, so the count is
-/// absolute (the spine's own hashing subtracts out) and hash-aware (the
-/// other hash's delta must be zero: the legs hash under the wrap hash
-/// alone). This is the formula the campaign's wrap-side prediction rides
-/// on: paths per ROUND plus the small prep trees, not per table per group.
-#[test]
-fn the_batched_query_census_matches_the_closed_form() {
-    let e = real_batched_epoch_with(super::proof_fixture::fixture_options());
-    let spine = batched_epoch_program(&e);
-    let full = batched_epoch_program_with(&e, true, false);
-    let count = |p: &LfmProgram, keccak: bool| -> usize {
-        p.instrs
-            .iter()
-            .filter(|i| match i {
-                super::instr::Instr::KeccakF(_) => keccak,
-                super::instr::Instr::Blake3(_) => !keccak,
-                _ => false,
-            })
-            .count()
-    };
-    let hash = super::edsl::WrapHash::production();
-    let per_query =
-        super::batched_epoch_verify::batched_query_permutations_for(&e.shape, &e.fri_params, hash);
-    let is_keccak = matches!(hash, super::edsl::WrapHash::Keccak);
-    let wrap_delta = count(&full, is_keccak) - count(&spine, is_keccak);
-    let other_delta = count(&full, !is_keccak) - count(&spine, !is_keccak);
-    assert_eq!(
-        wrap_delta,
-        e.proof.queries.len() * per_query,
-        "the legs' wrap-hash permutations must be exactly the census closed form"
-    );
-    assert_eq!(other_delta, 0, "the legs hash under the wrap hash alone");
-    eprintln!(
-        "batched query census: {per_query} wrap permutations/query over {} tables",
-        e.proof.tables.len()
-    );
-}
-
-/// ★★ The CARVED emitter gates — D1's emitted side, on the continuation
-/// batched format (the L2G main matrix carved standalone). One test, five
-/// arms: the assembled verifier RUNS on an honest carved epoch; the census
-/// closed form and the schema words match the carved program exactly
-/// (structural); a tampered carved ROOT, opening VALUE and SIBLING are each
-/// unprovable; and the wrong-reduction control fires on the carved walk
-/// (verdict condition 4's emitted side), discrimination checked, not hoped
-/// for.
-#[test]
-fn the_assembled_carved_batched_epoch_verifier_runs() {
-    let e = real_batched_epoch_carved_from(
-        super::proof_fixture::fixture_options(),
-        EpochInputs::from_env(),
-    );
-    let c = e
-        .shape
-        .carved_main
-        .expect("the harness carved the L2G table");
-    assert_eq!(
-        c.table,
-        e.proof.tables.len() - 1,
-        "the carve is the L2G bookend, the last table"
-    );
-    let h_carved = e.shape.heights[c.table];
-    let h_max_fri = e.shape.h_max();
-
-    let program = batched_epoch_program_with(&e, true, false);
-    let mut arenas = batched_epoch_arenas(&e);
-    arenas.push(super::epoch_verify_tests::batched_opening_arena(&e));
-    arenas.push(super::epoch_verify_tests::batched_fri_arena(&e));
-    execute(&program, &arenas, &crate::hash_pin::BLOCK_HASHER)
-        .expect("an honest carved batched epoch must run end to end");
-
-    // Structural: the census closed form and the schema words, on the CARVED
-    // shape — the same absolute guards the uncarved program carries.
-    let spine = batched_epoch_program_with(&e, false, false);
-    let count = |p: &LfmProgram, keccak: bool| -> usize {
-        p.instrs
-            .iter()
-            .filter(|i| match i {
-                super::instr::Instr::KeccakF(_) => keccak,
-                super::instr::Instr::Blake3(_) => !keccak,
-                _ => false,
-            })
-            .count()
-    };
-    let hash = super::edsl::WrapHash::production();
-    let per_query =
-        super::batched_epoch_verify::batched_query_permutations_for(&e.shape, &e.fri_params, hash);
-    let is_keccak = matches!(hash, super::edsl::WrapHash::Keccak);
-    assert_eq!(
-        count(&program, is_keccak) - count(&spine, is_keccak),
-        e.proof.queries.len() * per_query,
-        "the carved legs' wrap-hash permutations must be exactly the closed form"
-    );
-    for with_legs in [false, true] {
-        let p = batched_epoch_program_with(&e, with_legs, false);
-        let declared: usize = p.arena_schema.lens.iter().map(|l| *l as usize).sum();
-        assert_eq!(
-            declared,
-            expected_batched_arena_words(&e, with_legs),
-            "the carved program's schema must be exactly the shape's words"
-        );
-    }
-
-    // Tamper: the carved ROOT — its arena is index 2, right after the
-    // statement and the prep roots (declaration order = absorb order).
-    let open_idx = arenas.len() - 2;
-    let mut tampered = arenas.clone();
-    tampered[2][0] = base_word(FE::from(999_999u64));
-    assert!(
-        execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-        "a tampered carved root must not verify"
-    );
-
-    // Tamper: the carved opening VALUE and a carved path SIBLING — the
-    // carved block sits after the prep openings in the opening arena.
-    let mut off = 0usize;
-    for &(h, w) in &e.shape.prep.dims {
-        off += 2 * w + 2 * (h - 1);
-    }
-    let mut tampered = arenas.clone();
-    tampered[open_idx][off] = base_word(FE::from(999_999u64));
-    assert!(
-        execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-        "a tampered carved opening value must not verify"
-    );
-    let mut tampered = arenas.clone();
-    tampered[open_idx][off + 2 * c.width] = base_word(FE::from(999_999u64));
-    assert!(
-        execute(&program, &tampered, &crate::hash_pin::BLOCK_HASHER).is_err(),
-        "a tampered carved sibling must not verify"
-    );
-
-    // The wrong-reduction control, on the carved walk specifically.
-    if h_carved < h_max_fri {
-        let discriminates = e.challenges.fri.iotas.iter().any(|&iota| {
-            (iota >> (h_max_fri - h_carved)) != (iota & ((1usize << (h_carved - 1)) - 1))
-        });
-        if discriminates {
-            let wrong = batched_epoch_program_with(&e, true, true);
-            assert!(
-                execute(&wrong, &arenas, &crate::hash_pin::BLOCK_HASHER).is_err(),
-                "the wrong reduction direction must not authenticate the carved walk"
-            );
-        } else {
-            eprintln!("carved wrong-reduction arm skipped: the drawn indices do not discriminate");
-        }
-    } else {
-        eprintln!("carved wrong-reduction arm skipped: the carved table is the tallest");
-    }
-}
-
-/// ★★ P2's from-proof gates, batched: a continuation proven BATCHED wraps
-/// from the bundle alone. Epoch 0 AND the FINAL epoch reconstruct through
-/// [`real_batched_epoch_from_continuation`] (production's complete batched
-/// verify inside the harvest is the acceptance gate), each emits its CARVED
-/// program, the arenas fill to the schema, and the assembled verifier RUNS.
-/// A tampered bundle is rejected by the constructor's production verify —
-/// the P1 gate pair, on the batched format.
-#[test]
-fn the_batched_from_proof_constructor_runs_a_continuation_epoch() {
-    let elf_bytes = super::proof_fixture::read_inner_elf();
-    let opts = super::proof_fixture::fixture_options();
-    let bundle = crate::continuation::prove_continuation_batched(
-        &elf_bytes,
-        &[],
-        super::proof_fixture::FIXTURE_EPOCH_LOG2,
-        &opts,
-    )
-    .expect("the fixture continuation must prove batched");
-    assert!(
-        bundle.num_epochs() >= 2,
-        "the fixture continuation must have a second (final) epoch"
-    );
-
-    for (epoch, name) in [(0usize, "genesis"), (bundle.num_epochs() - 1, "FINAL")] {
-        let e = real_batched_epoch_from_continuation(&opts, &elf_bytes, &bundle, epoch, None)
-            .unwrap_or_else(|err| panic!("epoch {epoch} ({name}) must reconstruct: {err}"));
-        assert!(
-            e.shape.carved_main.is_some(),
-            "a continuation batched epoch is carved"
-        );
-        let program = batched_epoch_program_with(&e, true, false);
-        let mut arenas = batched_epoch_arenas(&e);
-        arenas.push(super::epoch_verify_tests::batched_opening_arena(&e));
-        arenas.push(super::epoch_verify_tests::batched_fri_arena(&e));
-        execute(&program, &arenas, &crate::hash_pin::BLOCK_HASHER).unwrap_or_else(|err| {
-            panic!("epoch {epoch} ({name})'s carved program must run: {err:?}")
-        });
-        eprintln!(
-            "★ P2 GATE: the {name} epoch's carved program ran ({} instrs)",
-            program.instrs.len()
-        );
-    }
-
-    // Tamper: a corrupted bound reg_fini is rejected by the constructor's
-    // production verify, exactly as on the per-table path.
-    let mut tampered = bundle;
-    tampered.corrupt_epoch_reg_fini_for_tests(1);
-    let err = match real_batched_epoch_from_continuation(&opts, &elf_bytes, &tampered, 1, None) {
-        Err(e) => e,
-        Ok(_) => panic!("a corrupted reg_fini must not reconstruct"),
-    };
-    assert!(
-        err.contains("rejects"),
-        "the corruption is caught by the production verify inside the constructor: {err}"
     );
 }
